backend/requirements.txt
Flask
flask-cors
requests
python-dotenv
gunicorn
google-generativeai
google-cloud-texttospeech